requirements.txt
APScheduler==3.6.1
Click==7.0
Django==2.2.6
django-filter==2.2.0
djangorestframework==3.10.3
dlib==19.18.0
face-recognition==1.2.3
face-recognition-models==0.3.0
Markdown==3.1.1
mysqlclient==1.4.4
numpy==1.17.3
Pillow==6.2.0
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
tzlocal==2.0.0
